Move multibinding test to BindingModuleMultibindingSetTest
